Split Effect -and EagerEffectScope tests
