+++ src/test/java/arrow/RewriteEffectScopeTest.java
@@ -8,7 +8,7 @@ import org.openrewrite.test.RewriteTest;
 
 import static org.openrewrite.kotlin.Assertions.kotlin;
 
-class RaiseRecipeTest implements RewriteTest {
+class RewriteEffectScopeTest implements RewriteTest {
     @Override
     public void defaults(RecipeSpec spec) {
         spec.recipe(
@@ -120,82 +120,6 @@ class RaiseRecipeTest implements RewriteTest {
 
               suspend fun Raise<Int>.test(): Unit =
                 ensure(false) { -1 }
-              """
-          )
-        );
-    }
-
-    @Test
-    void eagerEffectScopeParameter() {
-        rewriteRun(
-          kotlin(
-            """
-              package com.yourorg
-                            
-              import arrow.core.continuations.EagerEffectScope
-
-              fun test(scope: EagerEffectScope<String>): Int {
-                return scope.shift("failure")
-              }
-              """,
-            """
-              package com.yourorg
-                            
-              import arrow.core.raise.Raise
-
-              fun test(scope: Raise<String>): Int {
-                return scope.raise("failure")
-              }
-              """
-          )
-        );
-    }
-
-    @Test
-    void eagerEffectScopeReceiver() {
-        rewriteRun(
-          kotlin(
-            """
-              package com.yourorg
-                            
-              import arrow.core.continuations.EagerEffectScope
-
-              fun EagerEffectScope<String>.test(): Int {
-                return shift("failure")
-              }
-              """,
-            """
-              package com.yourorg
-                            
-              import arrow.core.raise.Raise
-
-              fun Raise<String>.test(): Int {
-                return raise("failure")
-              }
-              """
-          )
-        );
-    }
-
-    @Test
-    void eagerEffectScopeReceiverExpression() {
-        rewriteRun(
-          kotlin(
-            """
-              package com.yourorg
-                            
-              import arrow.core.continuations.EagerEffectScope
-
-              fun EagerEffectScope<String>.test(): Int =
-                shift("failure")
-              """,
-            """
-              package com.yourorg
-                            
-              import arrow.core.raise.Raise
-
-              fun Raise<String>.test(): Int =
-                raise("failure")
               """
           )
         );
